+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import type { Metadata, Viewport } from "next";
 import "@/styles/global.css";
 
 import { UserProvider } from "@/contexts/UserContext";
-import { ThemeProvider } from "@/components/core/theme-provider/ThemeProvider";
+import { ThemeProvider } from "@/components/theme-provider/ThemeProvider";
 
 export const viewport = {
   width: "device-width",
