+++ src/components/dashboard/layout/MobileNav.tsx
@@ -15,7 +15,7 @@ import { CaretUpDown as CaretUpDownIcon } from "@phosphor-icons/react/dist/ssr/C
 import type { NavItemConfig } from "@/types/nav";
 import { paths } from "@/paths";
 import { isNavItemActive } from "@/lib/is-nav-item-active";
-import { Logo } from "@/components/core/Logo";
+import { Logo } from "@/components/logo/Logo";
 
 import { logger } from "@/lib/default-logger";
 import { useUser } from "@/hooks/useUser";
